refactor(web): extract JSON feature boundary
