+++ examples/web/json.html
@@ -922,6 +922,6 @@
 
   </div>
 
-  <script type="module" src="/src/json-editor.ts"></script>
+  <script type="module" src="/src/entries/json.ts"></script>
 </body>
 </html>
