+++ examples/web/src/entries/json.ts
@@ -1,0 +1,3 @@
+import { mountJson } from '../features/json/browser/mount';
+
+mountJson();
