+++ examples/web/src/features/json/browser/mount.ts
@@ -1,0 +1,5 @@
+import { mountJsonEditor } from './editor';
+
+export function mountJson(): void {
+  mountJsonEditor();
+}
