+++ examples/web/src/shared/decoration-overlay.ts
